+++ test/screens/qr_code_generator_test.dart
@@ -7,7 +7,7 @@ import 'package:mobile_test/src/customWidgets/custom_countdown_timer.dart';
 import 'package:mobile_test/src/screens/qr_code_generator.dart';
 import 'package:qr_flutter/qr_flutter.dart';
 
-import '../mockSetUp/test_setup.dart';
+import '../mock_set_up/test_setup.dart';
 import '../model/seed_mock.dart';
 import 'pump_widget_mock_setup.dart';
